fixed warnings.

+++ src/final_state_generation_method.c
@@ -310,5 +310,5 @@ static RET_VAL _FreeFinalStateInternal( FINAL_STATE_GENARATOR_INTERNAL *internal
 }
 
 static int _CompareLogicalSpecies( CADDR_T species1, CADDR_T species2 ) {
-    return (int)species1 - (int)species2;
+  return (species1 - species2);
 }
